HBASE-7808 Refactor Store to use HRegionFileSystem
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.95@1460612 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/Store.java
@@ -146,7 +146,7 @@ public interface Store extends HeapSize, StoreConfigInformation {
    * @param includeMVCCReadpoint whether we should out the MVCC readpoint
    * @return Writer for a new StoreFile in the tmp dir.
    */
-  public StoreFile.Writer createWriterInTmp(int maxKeyCount, Compression.Algorithm compression,
+  public StoreFile.Writer createWriterInTmp(long maxKeyCount, Compression.Algorithm compression,
       boolean isCompaction, boolean includeMVCCReadpoint) throws IOException;
 
   // Compaction oriented methods
